pyproject.toml
[tool.poetry]
name = "salesforce-extractor"
version = "0.1.0"
description = "Salesforce extractor"
authors = ["GOTO Satoru <satoru.goto@cognite.com>"]

[tool.poetry.dependencies]
python = "^3.8"
cognite-extractor-utils = "^2.2.0"
simple_salesforce = "^1.12.2"

[tool.poetry.dev-dependencies]
isort = "^5.9.3"
flake8 = "^3.9.2"
mypy = "^0.910"
black = "^20.8b1"
simple_salesforce = "^1.12.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
salesforce_extractor = "salesforce_extractor.extractor:main"
